Refactor ABI and support packing events
I wanted to be able to integration test vent and noticed there was no `PackEvent`. When I went to add this I saw that ABI had become a bit overgrown and hard to follow so I factored it out into a few different files. Outside the reorganisation:

- Implemented `PackEvent`
- Added tests for event packing
- Reduced public API service by making `Pack*` and `Unpack*` work with variadic args or a struct combing the functionality of `PackIntoStruct` (I didn't want us maintain `PackEventStruct` `UnpackEventStruct` and all the other variants.
- Supported packing of uint8, int8, and big.Int (we were unpacking big ints, but didn't support packing them)
- Made `pack` elide indexed fields (i.e. fields that go into Log topics)
in the same way that `unpack` does
