+++ execution/evm/abi/abi_test.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"testing"
 
-	hex "github.com/tmthrgd/go-hex"
+	"github.com/tmthrgd/go-hex"
 
 	"github.com/hyperledger/burrow/logging"
 	"github.com/stretchr/testify/assert"
